# perf: use simplified internal string