--- a/crates/examples/examples/sheet.rs
+++ b/crates/examples/examples/sheet.rs
@@ -1,5 +1,6 @@
-use dev_utils::get_mem_usage;
+use dev_utils::{get_mem_usage, ByteSize};
 use examples::sheet::init_large_sheet;
+use loro::ID;
 
 pub fn main() {
     dev_utils::setup_test_log();
@@ -10,53 +11,53 @@ pub fn main() {
     println!("Allocated bytes for 10M cells spreadsheet: {}", allocated);
     println!("Has history cache: {}", doc.has_history_cache());
     examples::utils::bench_fast_snapshot(&doc);
-    // doc.checkout(&ID::new(doc.peer_id(), 100).into()).unwrap();
-    // println!(
-    //     "Has history cache after checkout: {}",
-    //     doc.has_history_cache()
-    // );
+    doc.checkout(&ID::new(doc.peer_id(), 100).into()).unwrap();
+    println!(
+        "Has history cache after checkout: {}",
+        doc.has_history_cache()
+    );
 
-    // doc.checkout_to_latest();
-    // let after_checkout = get_mem_usage();
-    // println!("Allocated bytes after checkout: {}", after_checkout);
+    doc.checkout_to_latest();
+    let after_checkout = get_mem_usage();
+    println!("Allocated bytes after checkout: {}", after_checkout);
 
-    // doc.free_diff_calculator();
-    // let after_free_diff_calculator = get_mem_usage();
-    // println!(
-    //     "Allocated bytes after freeing diff calculator: {}",
-    //     after_free_diff_calculator
-    // );
+    doc.free_diff_calculator();
+    let after_free_diff_calculator = get_mem_usage();
+    println!(
+        "Allocated bytes after freeing diff calculator: {}",
+        after_free_diff_calculator
+    );
 
-    // println!(
-    //     "Diff calculator size: {}",
-    //     after_checkout - after_free_diff_calculator
-    // );
+    println!(
+        "Diff calculator size: {}",
+        after_checkout - after_free_diff_calculator
+    );
 
-    // doc.free_history_cache();
-    // let after_free_history_cache = get_mem_usage();
-    // println!(
-    //     "Allocated bytes after free history cache: {}",
-    //     after_free_history_cache
-    // );
+    doc.free_history_cache();
+    let after_free_history_cache = get_mem_usage();
+    println!(
+        "Allocated bytes after free history cache: {}",
+        after_free_history_cache
+    );
 
-    // println!(
-    //     "History cache size: {}",
-    //     after_free_diff_calculator - after_free_history_cache
-    // );
+    println!(
+        "History cache size: {}",
+        after_free_diff_calculator - after_free_history_cache
+    );
 
-    // doc.compact_change_store();
-    // let after_compact_change_store = get_mem_usage();
-    // println!(
-    //     "Allocated bytes after compact change store: {}",
-    //     after_compact_change_store
-    // );
-    // println!(
-    //     "Shrink change store size: {}",
-    //     after_free_history_cache - after_compact_change_store
-    // );
+    doc.compact_change_store();
+    let after_compact_change_store = get_mem_usage();
+    println!(
+        "Allocated bytes after compact change store: {}",
+        after_compact_change_store
+    );
+    println!(
+        "Shrink change store size: {}",
+        after_free_history_cache - after_compact_change_store
+    );
 
-    // println!(
-    //     "ChangeStore size: {}",
-    //     ByteSize(doc.with_oplog(|log| log.change_store_kv_size()))
-    // );
+    println!(
+        "ChangeStore size: {}",
+        ByteSize(doc.with_oplog(|log| log.change_store_kv_size()))
+    );
 }