--- a/crates/loro-common/Cargo.toml
+++ b/crates/loro-common/Cargo.toml
@@ -21,7 +21,6 @@ thiserror = "1.0.43"
 wasm-bindgen = { version = "=0.2.92", optional = true }
 fxhash = "0.2.1"
 enum-as-inner = "0.6.0"
-string_cache = "0.8"
 arbitrary = { version = "1.3.0", features = ["derive"] }
 js-sys = { version = "0.3.60", optional = true }
 serde_columnar = { workspace = true }