In the old implementation of InternalString, performance would degrade significantly when dealing with a large number of different InternalStrings.

In the following example

```rust
use std::time::Instant;

use dev_utils::get_mem_usage;
use loro::LoroDoc;

fn main() {
    let mut init = Instant::now();
    {
        let start = Instant::now();
        let doc = LoroDoc::new();
        let map = doc.get_map("map");
        for i in 0..1000 {
            for j in 0..1000 {
                let key = format!("0000{}:0000{}", i, j);
                map.insert(&key, i + j).unwrap();
            }
        }

        println!("LargeMap Init Time {:?}", start.elapsed());
        let start = Instant::now();
        let bytes = doc.export(loro::ExportMode::Snapshot).unwrap();
        println!("LargeMap Export Time {:?}", start.elapsed());
        let start = Instant::now();
        let new_doc = LoroDoc::new();
        new_doc.import(&bytes).unwrap();
        println!("LargeMap Import Time {:?}", start.elapsed());
        println!("Mem {:?}", get_mem_usage());
        init = Instant::now();
    }
    println!("Drop Time {:?}", init.elapsed());
    println!("Mem {:?}", get_mem_usage())
}
```

**Before this PR**

```
LargeMap Init Time 15.036099875s
LargeMap Export Time 1.127982417s
LargeMap Import Time 33.742875ms
Mem 271.72 MB
Drop Time 11.854144s
Mem 65.27 KB
```

The majority of time is spent on the old InternalString's dropping and creation. Because it became very slow to create new InternalStrings when we retained 1M InternalStrings in the memory.

**After this PR**

```
LargeMap Init Time 1.276956416s
LargeMap Export Time 1.207581209s
LargeMap Import Time 111.589458ms
Mem 282.05 MB
Drop Time 405.822667ms
Mem 2.47 KB
```